+++ app/app.ts
@@ -1,15 +1,16 @@
-/// <reference path="player.ts" />
-/// <reference path="game.ts" />
+import { Player } from './player';
+import { Game } from './game';
+import * as Helpers from './utility';
 
 let newGame: Game;
 
 // add click handler to the start game button
 document.getElementById('startGame')!.addEventListener('click', () => {
   const player: Player = new Player();
-  player.name = Utility.getInputValue('playername');
+  player.name = Helpers.getValue('playername');
 
-  const problemCount: number = Number(Utility.getInputValue('problemCount'));
-  const factor: number = Number(Utility.getInputValue('factor'));
+  const problemCount: number = Number(Helpers.getValue('problemCount'));
+  const factor: number = Number(Helpers.getValue('factor'));
 
   newGame = new Game(player, problemCount, factor);
   newGame.displayGame();
