+++ app/person.ts
@@ -1,4 +1,4 @@
-interface Person {
+export interface Person {
   name: string;
   age?: number;
   formatName: () => string;
