+++ app/player.ts
@@ -1,6 +1,6 @@
-/// <reference path="person.ts" />
+import { Person } from './person';
 
-class Player implements Person {
+export class Player implements Person {
   name: string;
   age: number;
   highScore: number;
