+++ app/result.ts
@@ -1,4 +1,4 @@
-interface Result {
+export interface Result {
   playerName: string;
   score: number;
   problemCount: number;
