+++ app/tsconfig.json
@@ -2,7 +2,9 @@
   "extends": "../tsconfig.base",
   "compilerOptions": {
     "removeComments": true,
-    "outFile": "../js/app.js"
+    "module": "commonjs",
+    "moduleResolution": "node",
+    "traceResolution": true
   },
   "files": [
     "./app.ts"
