+++ index.html
@@ -55,7 +55,7 @@
     <div class="col-sm-offset-2 col-sm-10" id="messages"></div>
     
     <!-- add script tags here -->
-    <script src="./js/app.js"></script>
+    <script src="bundle.js"></script>
 
 </body>
 </html>
