+++ js/app.js
@@ -1,87 +1,15 @@
-var Player = (function () {
-    function Player() {
-    }
-    Player.prototype.formatName = function () {
-        return this.name.toUpperCase();
-    };
-    return Player;
-}());
-var Utility = (function () {
-    function Utility() {
-    }
-    Utility.getInputValue = function (elementID) {
-        var inputElement = document.getElementById(elementID);
-        return inputElement.value;
-    };
-    return Utility;
-}());
-var Scoreboard = (function () {
-    function Scoreboard() {
-        this.results = [];
-    }
-    Scoreboard.prototype.addResult = function (newResult) {
-        this.results.push(newResult);
-    };
-    Scoreboard.prototype.updateScoreboard = function () {
-        var output = '<h2>Scoreboard</h2>';
-        for (var index = 0; index < this.results.length; index++) {
-            var result = this.results[index];
-            output += '<h4>';
-            output += result.playerName + ': ' + result.score + '/' + result.problemCount + ' for factor ' + result.factor;
-            output += '</h4>';
-        }
-        var scoresElement = document.getElementById('scores');
-        scoresElement.innerHTML = output;
-    };
-    return Scoreboard;
-}());
-var Game = (function () {
-    function Game(player, problemCount, factor) {
-        this.player = player;
-        this.problemCount = problemCount;
-        this.factor = factor;
-        this.scoreboard = new Scoreboard();
-    }
-    Game.prototype.displayGame = function () {
-        var gameForm = '';
-        for (var i = 1; i <= this.problemCount; i++) {
-            gameForm += '<div class="form-group">';
-            gameForm += '<label for="answer' + i + '" class="col-sm-2 control-label">';
-            gameForm += String(this.factor) + ' x ' + i + ' = </label>';
-            gameForm += '<div class="col-sm-1"><input type="text" class="form-control" id="answer' + i + '" size="5" /></div>';
-            gameForm += '</div>';
-        }
-        var gameElement = document.getElementById('game');
-        gameElement.innerHTML = gameForm;
-        document.getElementById('calculate').removeAttribute('disabled');
-    };
-    Game.prototype.calculateScore = function () {
-        var score = 0;
-        for (var i = 1; i <= this.problemCount; i++) {
-            var answer = Number(Utility.getInputValue('answer' + i));
-            if (i * this.factor === answer) {
-                score++;
-            }
-        }
-        var result = {
-            playerName: this.player.name,
-            score: score,
-            problemCount: this.problemCount,
-            factor: this.factor
-        };
-        this.scoreboard.addResult(result);
-        this.scoreboard.updateScoreboard();
-        document.getElementById('calculate').setAttribute('disabled', 'true');
-    };
-    return Game;
-}());
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+var player_1 = require("./player");
+var game_1 = require("./game");
+var Helpers = require("./utility");
 var newGame;
 document.getElementById('startGame').addEventListener('click', function () {
-    var player = new Player();
-    player.name = Utility.getInputValue('playername');
-    var problemCount = Number(Utility.getInputValue('problemCount'));
-    var factor = Number(Utility.getInputValue('factor'));
-    newGame = new Game(player, problemCount, factor);
+    var player = new player_1.Player();
+    player.name = Helpers.getValue('playername');
+    var problemCount = Number(Helpers.getValue('problemCount'));
+    var factor = Number(Helpers.getValue('factor'));
+    newGame = new game_1.Game(player, problemCount, factor);
     newGame.displayGame();
 });
 document.getElementById('calculate').addEventListener('click', function () {
