+++ js/game.js
@@ -1,0 +1,47 @@
+"use strict";
+Object.defineProperty(exports, "__esModule", { value: true });
+exports.Game = void 0;
+var utility_1 = require("./utility");
+var scoreboard_1 = require("./scoreboard");
+var Game = (function () {
+    function Game(player, problemCount, factor) {
+        this.player = player;
+        this.problemCount = problemCount;
+        this.factor = factor;
+        this.scoreboard = new scoreboard_1.Scoreboard();
+    }
+    Game.prototype.displayGame = function () {
+        var gameForm = '';
+        for (var i = 1; i <= this.problemCount; i++) {
+            gameForm += '<div class="form-group">';
+            gameForm += '<label for="answer' + i + '" class="col-sm-2 control-label">';
+            gameForm += String(this.factor) + ' x ' + i + ' = </label>';
+            gameForm += '<div class="col-sm-1"><input type="text" class="form-control" id="answer' + i + '" size="5" /></div>';
+            gameForm += '</div>';
+        }
+        var gameElement = document.getElementById('game');
+        gameElement.innerHTML = gameForm;
+        document.getElementById('calculate').removeAttribute('disabled');
+    };
+    Game.prototype.calculateScore = function () {
+        var score = 0;
+        for (var i = 1; i <= this.problemCount; i++) {
+            var answer = Number(utility_1.getValue('answer' + i));
+            if (i * this.factor === answer) {
+                score++;
+            }
+        }
+        var result = {
+            playerName: this.player.name,
+            score: score,
+            problemCount: this.problemCount,
+            factor: this.factor
+        };
+        this.scoreboard.addResult(result);
+        this.scoreboard.updateScoreboard();
+        document.getElementById('calculate').setAttribute('disabled', 'true');
+    };
+    return Game;
+}());
+exports.Game = Game;
+//# sourceMappingURL=game.js.map
