+++ webpack.config.js
@@ -1,22 +1,22 @@
 module.exports = {
-  entry: './js/app.js',
-  // devtool: 'inline-source-map',
-  // mode: 'development',
-  // module: {
-  //   rules: [
-  //     {
-  //       test: /\.tsx?$/,
-  //       use: 'ts-loader',
-  //       exclude: /node_modules/
-  //     }
-  //   ]
-  // },
-  // resolve: {
-  //   extensions: [ '.tsx', '.ts', '.js' ]
-  // },
-  // output: {
-  //   filename: 'bundle.js'
-  // },
+  entry: './app/app.ts',
+  devtool: 'inline-source-map',
+  mode: 'development',
+  module: {
+    rules: [
+      {
+        test: /\.tsx?$/,
+        use: 'ts-loader',
+        exclude: /node_modules/
+      }
+    ]
+  },
+  resolve: {
+    extensions: [ '.tsx', '.ts', '.js' ]
+  },
+  output: {
+    filename: 'bundle.js'
+  },
   devServer: {
     inline: false
   }
